Module3, Assignment Information
"Compare results between set #1 vs set #2 by discussing the differences between the two sets." (Module 3, Assignment Question)

For my answer, the differences lie in the range(s), as the first data set were primarily lower data and the second contains higher integers.  The range was half of one another, with the first being 2-10, and the second 12-20.  Both carries the same (sd)/standard deviation, so there is no change in them both being almost even, but slightly skewed to the right.  (The mode for set1=2, set2=12).